Add reset button and set static grid width
Created resetButton function to add a rest button to the webpage,
when clicked the user is prompted to enter a number to set the new grid resolution.
Added event listener to check when reset button is clicked.
Updated createGrid function to remove old grid when reset button is clicked.
Updated creatGrid to size the squares so that the grid remains the same size regardless of number of squares.